[수정] 태그 검색 후 모아보기 화면 삭제 후 병합
