+++ SOOUM/SOOUM/Presentations/Main/Tags/Search/Views/SearchTextFieldView+Rx.swift
@@ -1,0 +1,16 @@
+//
+//  SearchTextFieldView+Rx.swift
+//  SOOUM
+//
+//  Created by 오현식 on 1/2/26.
+//
+
+import RxCocoa
+import RxSwift
+
+extension Reactive where Base: SearchTextFieldView {
+    
+    var text: ControlProperty<String?> {
+        self.base.textField.rx.text
+    }
+}
